Fix immediate-door timing and clarify elapsed math

+++ firmware/master_f4/src/master_controller.c
@@ -4,7 +4,7 @@
 #define DIRECTION_MISMATCH_PENALTY 2U
 
 static uint32_t elapsed_ms(uint32_t now_ms, uint32_t previous_ms) {
-    return now_ms - previous_ms;
+    return (uint32_t)(now_ms - previous_ms);
 }
 
 static uint8_t clamp_floor(uint8_t floor) {
